feat: removed firebase calls from frontend, fixed events page
