--- app/(default)/events/EventCard.css
@@ -1,215 +1,0 @@
-h2 {
-  font-family: 'Arial Narrow Bold', sans-serif;
-  margin-top: 60px;
-  padding-bottom: 30px;
-  font-size: 2.5em;
-  color: #00c853;
-  text-align: center;
-}
-
-.event-cards-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-}
-
-.event-card {
-  width: 300px;
-  height: 400px;
-  perspective: 1000px;
-  margin: 20px;
-}
-
-.event-card-inner {
-  width: 100%;
-  height: 100%;
-  position: relative;
-  transform-style: preserve-3d;
-  transition: transform 0.8s;
-}
-
-.event-card:hover .event-card-inner {
-  transform: rotateY(180deg);
-}
-
-.event-card-front,
-.event-card-back {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  backface-visibility: hidden;
-  background-color: black;
-  color: white;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  border: 2px solid #00c853;
-}
-
-.event-card-front {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.event-card-back {
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  transform: rotateY(180deg);
-}
-
-.event-card-back img {
-  width: 100%;
-  height: auto;
-  max-height: 150px;
-  opacity: 0.2;
-  object-fit: cover;
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 0;
-}
-
-.event-card-back p {
-  z-index: 1;
-  position: relative;
-  color: white;
-  font-size: 1.2em;
-  text-align: center;
-}
-
-.event-poster-container {
-  position: relative;
-  width: 280px;
-  padding-bottom: 75%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.event-poster {
-  position: absolute;
-  top: -40px;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 10px;
-}
-
-.event-content-container {
-  text-align: center;
-}
-
-.event-content-container h3 {
-  margin-bottom: 0;
-  font-size: 1.4em;
-}
-
-.event-content-container p {
-  margin-top: 0px;
-  margin-bottom: 25px;
-  font-size: 1em;
-}
-
-/* Updated button styles */
-.toggle-form-button,
-.toggle-edit-button,
-.submit-event-button,
-.update-event-button,
-.remove-event-button {
-  margin-top: 70px !important;
-  background-color: black;
-  color: yellow;
-  border: 2px solid yellow;
-  border-radius: 5px;
-  padding: 10px 20px;
-  font-size: 1em;
-  cursor: pointer;
-  transition: background-color 0.3s, color 0.3s;
-  margin: 10px;
-}
-
-.toggle-form-button:hover,
-.toggle-edit-button:hover,
-.submit-event-button:hover,
-.update-event-button:hover,
-.remove-event-button:hover {
-  background-color: yellow;
-  color: black;
-}
-
-form {
-  background-color: black;
-  color: white;
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
-  border: 2px solid yellow;
-  width: 80%;
-  margin: 40px auto;
-}
-
-form label {
-  color: yellow;
-  font-weight: bold;
-  display: block;
-  margin-bottom: 10px;
-}
-
-form input,
-form textarea,
-form select {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 20px;
-  border: 2px solid yellow;
-  border-radius: 5px;
-  background-color: black !important;
-  color: white !important;
-  font-size: 1em;
-}
-
-form input::placeholder,
-form textarea::placeholder {
-  color: grey !important;
-  opacity: 1;
-}
-
-form input:focus,
-form textarea:focus,
-form select:focus {
-  border-color: yellow;
-  outline: none;
-}
-
-input[type="text"],
-input[type="date"],
-textarea {
-  background-color: black !important;
-  color: white !important;
-}
-
-input[type="text"]::placeholder,
-input[type="date"]::placeholder,
-textarea::placeholder {
-  color: grey !important;
-  opacity: 1;
-}
-
-form button {
-  margin-top: 20px;
-}
-
-section {
-  margin-bottom: 40px;
-}
-
-.upcoming-events,
-.ongoing-events,
-.past-events {
-  margin-top: 60px;
-}
