+++ package.json
@@ -47,7 +47,8 @@
     "sharp": "^0.33.5",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "uuid": "^11.0.3"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.7",
